pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "python-codicefiscale"
description = "encode / decode italian fiscal codes - codifica / decodifica del Codice Fiscale italiano."
authors = [
    { name = "Fabio Caccamo", email = "fabio.caccamo@gmail.com" },
]
keywords = [
    "codicefiscale",
    "codice",
    "fiscale",
    "cf",
    "fiscal code",
    "italia",
    "italy",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Topic :: Software Development :: Build Tools",
]
dependencies = [
    "cryptography>=42.0.8",
    "fastapi>=0.116.1",
    "functions-framework>=3.9.2",
    "jinja2>=3.1.6",
    "pyjwt[crypto]>=2.10.1",
    "python-dateutil >= 2.8, < 2.10",
    "python-dotenv>=1.1.1",
    "python-fsutil >= 0.10.0, < 1.0.0",
    "python-multipart>=0.0.20",
    "python-slugify >= 7.0.0, < 9.0.0",
    "uvicorn>=0.35.0",
]

dynamic = ["version"]
maintainers = [
    { name = "Fabio Caccamo", email = "fabio.caccamo@gmail.com" },
]

[project.optional-dependencies]
api = [
    "fastapi >= 0.100.0, < 1.0.0",
    "uvicorn >= 0.20.0, < 1.0.0",
    "pyjwt[crypto] >= 2.8.0, < 3.0.0",
    "cryptography >= 41.0.0, < 43.0.0",
    "python-dotenv >= 1.0.0, < 2.0.0",
    "jinja2 >= 3.1.0, < 4.0.0",
    "python-multipart >= 0.0.6, < 1.0.0",
]

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.license]
file = "LICENSE.txt"
content-type = "text/plain"

[project.urls]
Homepage = "https://github.com/fabiocaccamo/python-codicefiscale"
Download = "https://github.com/fabiocaccamo/python-codicefiscale/releases"
Documentation = "https://github.com/fabiocaccamo/python-codicefiscale#readme"
Issues = "https://github.com/fabiocaccamo/python-codicefiscale/issues"
Funding = "https://github.com/sponsors/fabiocaccamo/"
Twitter = "https://twitter.com/fabiocaccamo"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | venv
)/
'''

[tool.mypy]
files = ["codicefiscale"]
ignore_missing_imports = true
install_types = true
non_interactive = true
strict = true

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py310"
src = ["codicefiscale", "tests"]

[tool.ruff.lint]
select = [
    "B",    # flake8-bugbear
    "B9",   # flake8-bugbear opinionated
    "C",    # mccabe
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "W",    # pycodestyle warnings
    "UP",   # pyupgrade
    "RUF",  # ruff-specific rules
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.isort]
known-first-party = ["codicefiscale"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.setuptools]
packages = ["codicefiscale"]

[tool.setuptools.dynamic.version]
attr = "codicefiscale.metadata.__version__"

[dependency-groups]
dev = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "pytest==8.4.*",
    "pytest-cov==6.2.*",
    "ruff>=0.7.0",
    "uvicorn>=0.35.0",
]
